feat: Create a preview when you upload a file / Create new component for SeedEntryForm
