+++ client/src/pages/seedEntryPage/seedEntryPage.module.css
@@ -1,92 +1,6 @@
-.formContainer {
-  background-color: var(--color-background);
-  border-radius: 1em;
-  padding: 0.5em;
-}
-
-.form {
+.mainContainer {
   display: flex;
-  flex-direction: column;
-  align-items: center;
-  border: solid 0.3em var(--color-primary);
-  border-radius: 1em;
-  gap: 1em;
+  flex-direction: row;
+  justify-content: center;
   padding: 1em;
-  font-family: var(--text-font);
-  color: var(--color-primary);
-}
-
-.title {
-  font-family: var(--title-font);
-  font-size: 2em;
-  text-align: center;
-  margin-bottom: 1em;
-}
-
-.formLabel {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.formInput {
-  background-color: whitesmoke;
-  box-shadow: inset -1px 1px 3px 0 var(--color-hint);
-  height: 2em;
-  border-radius: 0.5em;
-  margin: 0.5em;
-  padding: 0.3em;
-  width: 30em;
-}
-
-.formTextArea {
-  background-color: whitesmoke;
-  box-shadow: inset -1px 1px 3px 0 var(--color-hint);
-  height: 2em;
-  border-radius: 0.5em;
-  margin: 0.5em;
-  padding: 0.3em;
-  width: 30em;
-  height: 20em;
-  resize: none;
-  ::placeholder {
-    font-family: var(--text-font);
-    font-size: 1.5em;
-    color: red;
-  }
-}
-
-.imageUploader {
-  height: 2em;
-  margin: 0.5em;
-  padding: 0.3em;
-}
-
-.submitButton {
-  background-color: var(--color-primary);
-  border: solid 0.1em var(--color-accent);
-  outline: solid 0.2em var(--color-primary);
-  border-radius: 1em;
-  margin: 1em;
-  padding: 0.5em 1.5em 0.5em 1.5em;
-  color: var(--color-accent);
-  font-family: var(--special-font);
-  font-size: 1.5em;
-}
-
-.text {
-  font-size: 0.8em;
-}
-
-.link {
-  font-size: 0.8em;
-  font-weight: bold;
-  &:hover {
-    text-decoration: underline;
-  }
-}
-
-span {
-  color: red;
-  font-size: 0.9em;
 }
